refactor(core): extract maintenance/dedup.ts — near-duplicate memory deactivation
